minimal working server
