+++ server/schema/config.js
@@ -1,0 +1,9 @@
+const dbConnection = {
+    host : '127.0.0.1',
+    user : 'root',
+    password : 'password',
+    database : 'dcs'
+ }
+
+module.exports.dbConnection = dbConnection
+
