--- server/schemas/resolver.js
@@ -1,9 +1,0 @@
-
-const resolvers = {
-
-	Query: {
-
-	}
-}
-
-module.exports.resolvers = resolvers
