+++ server/server.js
@@ -1,17 +1,13 @@
 const { ApolloServer } = require('apollo-server');
-const { typeDefs } = require('./schema/typedefs.js');
+const { typeDefs } = require('./schema/typedef.js');
 const { resolvers } = require('./schema/resolver.js');
 const { CatalogNext } = require('./schema/datasource.js')
+const { dbConnection } = require('./schema/config.js')
 
 // the database connetion details to be updated on deployment
 const knexConfig = {
   client: 'mysql',
-  connection: {
-    host : '127.0.0.1',
-    user : 'root',
-    password : 'password',
-    database : 'dcs'
-  }
+  connection: dbConnection
 };
 
 const db = new CatalogNext(knexConfig);
